try get aprove 4

+++ tasks/seq/Odintsov_M_CountingMismatchedCharactersStr/src/ops_seq.cpp
@@ -29,7 +29,7 @@ bool Odintsov_M_CountingMismatchedCharactersStr_seq::CountingCharacterSequential
 }
 bool Odintsov_M_CountingMismatchedCharactersStr_seq::CountingCharacterSequential::run() {
   internal_order_test();
-  int size = strlen(input[0]);
+  size_t size = strlen(input[0]);
   for (size_t i = 0; i < size; i++) {
     if (i < strlen(input[1])) {
       if (input[0][i] != input[1][i]) {
